+++ src/main/java/com/uniminuto/clinica/repository/RecetaRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RecetaRepository extends JpaRepository<Receta, Integer> {
+public interface RecetaRepository extends JpaRepository<Receta, Long> {
     List<Receta> findByCitaAndMedicamento(Cita cita, Medicamento medicamento);
-
 }
+
